backend/requirements.txt
fastapi
openapi
load_dotenv
uvicorn
slowapi
aiocache
redis